+++ internal/discover/dns/subdomain/helpers.go
@@ -3,6 +3,9 @@ package subdomain
 import (
 	// standard
 	"context"
+	"crypto/rand"
+	"fmt"
+	"math/big"
 	"net"
 )
 
@@ -24,6 +27,22 @@ func detectWildcardDNS(ctx context.Context, domain string, resolver *net.Resolve
 		return &wildcardDomain, nil
 	}
 
-	// If the error is NXDOMAIN or SERVFAIL, wildcard is NOT present
+	// If the error, domain didnt resolve so wildcard is NOT present
 	return nil, nil
+}
+
+// generateRandomSubdomain generates a high-entropy subdomain with only letters (16 characters).
+func generateRandomSubdomain(domain string) (string, error) {
+	letterBytes := "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+
+	randomString := make([]byte, 16)
+	for i := range randomString {
+		n, err := rand.Int(rand.Reader, big.NewInt(int64(len(letterBytes))))
+		if err != nil {
+			return "", err
+		}
+		randomString[i] = letterBytes[n.Int64()]
+	}
+
+	return fmt.Sprintf("%s.%s", string(randomString), domain), nil
 }
